Revert "Defines: SOCKET_ERROR - fix"
This reverts commit 131b4bb0a0ef529c384a455f722bf740c84e7cc8.

+++ Include/xLib/Core/Defines.h
@@ -666,14 +666,14 @@
 ///@{
 #if xENV_WIN
     #define xSOCKET_ERROR \
-        ( static_cast<ssize_t>(SOCKET_ERROR) )
+        ( SOCKET_ERROR )
         ///< socket native handle value "error"
     #define xSOCKET_HANDLE_INVALID \
         ( static_cast<socket_t>( INVALID_SOCKET ) )
         ///< socket native handle value "null"
 #elif xENV_UNIX
     #define xSOCKET_ERROR \
-        ( static_cast<ssize_t>(- 1) )
+        ( - 1 )
         ///< socket native handle value "error"
     #define xSOCKET_HANDLE_INVALID \
         ( static_cast<socket_t>( - 1 ) )
